library_management/requirements.txt
Django==4.2
djangorestframework
gunicorn
django_filter
django-cors-headers
